requirements.txt
channels==3.0.3
channels-redis==3.2.0
Django==3.2.4
psycopg2-binary==2.9.1
django-crispy-forms==1.12.0
djangorestframework==3.12.4
markdown==3.3.4
django_filter==2.4.0